Add multi-arch support for building the image
TARGETARCH env variable is used to specify for which architecture the
binary should be built.
TARGETARCH is in-box env variable for docker buildx, so nothing needs to
be sent from used side.
In case if pure docker is used - if TARGETARCH env variable is not sent,
amd64 binary will be built, similar to current behaviour.

Signed-off-by: Yulia Gaponenko <yulia.gaponenko1@de.ibm.com>